[quick_actions] Extract the Android implementation into a separate package (#4698)
