+++ packages/quick_actions/quick_actions_android/CHANGELOG.md
@@ -1,0 +1,3 @@
+## 0.6.0+9
+
+* Switches to a package-internal implementation of the platform interface.
