+++ packages/quick_actions/quick_actions_android/android/src/main/AndroidManifest.xml
