+++ packages/quick_actions/quick_actions_android/example/README.md
@@ -1,0 +1,8 @@
+# quick_actions_example
+
+Demonstrates how to use the quick_actions plugin.
+
+## Getting Started
+
+For help getting started with Flutter, view our online
+[documentation](https://flutter.dev/).
