+++ packages/quick_actions/quick_actions_android/example/android/app/src/main/AndroidManifest.xml
@@ -1,0 +1,22 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+          package="io.flutter.plugins.quickactionsexample">
+
+  <uses-permission android:name="android.permission.INTERNET"/>
+
+  <application android:label="quick_actions_example" android:icon="@mipmap/ic_launcher">
+      <meta-data
+          android:name="io.flutter.app.android.SplashScreenUntilFirstFrame"
+          android:value="true"/>
+      <activity android:name="io.flutter.embedding.android.FlutterActivity"
+          android:theme="@style/LaunchTheme"
+          android:configChanges="orientation|keyboardHidden|keyboard|screenSize|locale|layoutDirection"
+          android:hardwareAccelerated="true"
+          android:windowSoftInputMode="adjustResize">
+          <intent-filter>
+            <action android:name="android.intent.action.MAIN"/>
+            <category android:name="android.intent.category.LAUNCHER"/>
+          </intent-filter>
+      </activity>
+    <meta-data android:name="flutterEmbedding" android:value="2"/>
+  </application>
+</manifest>
