+++ packages/quick_actions/quick_actions_android/example/integration_test/quick_actions_test.dart
@@ -1,0 +1,24 @@
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'package:flutter_test/flutter_test.dart';
+import 'package:integration_test/integration_test.dart';
+import 'package:quick_actions_platform_interface/quick_actions_platform_interface.dart';
+
+void main() {
+  IntegrationTestWidgetsFlutterBinding.ensureInitialized();
+
+  testWidgets('Can set shortcuts', (WidgetTester tester) async {
+    final QuickActionsPlatform quickActions = QuickActionsPlatform.instance;
+    await quickActions.initialize((String value) {});
+
+    const ShortcutItem shortCutItem = ShortcutItem(
+      type: 'action_one',
+      localizedTitle: 'Action one',
+      icon: 'AppIcon',
+    );
+    expect(
+        quickActions.setShortcutItems(<ShortcutItem>[shortCutItem]), completes);
+  });
+}
